Fixed Javadocs on MetricSampler, extracted MetricSamplerOptions to outer class, added Javadocs to AbstractMetricSampler.
